# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Book Store API",
        "version": "0.1"
    },
    "paths": {
...
        "/api/auth/cms/login": {
            "post": {
                "tags": [
                    "Tài khoản CMS"
                ],
                "summary": "Đăng nhập CMS",
                "operationId": "cms_login",
                "parameters": [
                    {
                        "name": "X-localication",
                        "in": "header",
                        "description": "Ngôn ngữ",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "vi"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            },
                            "examples": {
                                "Examples": {
                                    "summary": "Examples",
                                    "value": {
                                        "email": "hungtmt1102@gmail.com",
                                        "password": "123123"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Success."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/cms/me": {
            "get": {
                "tags": [
                    "Tài khoản CMS"
                ],
                "summary": "Thông tin người dùng",
                "operationId": "cms_me",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Success."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/auth/cms/logout": {
            "get": {
                "tags": [
                    "Tài khoản CMS"
                ],
                "summary": "Đăng xuất",
                "operationId": "cms_logout",
                "parameters": [
                    {
                        "name": "language",
                        "in": "header",
                        "description": "Ngôn ngữ",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "vi"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Success."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/auth/cms/change-password": {
            "post": {
                "tags": [
                    "Tài khoản CMS"
                ],
                "summary": "Đổi mật khẩu",
                "operationId": "cms_change-password",
                "parameters": [
                    {
                        "name": "language",
                        "in": "header",
                        "description": "Ngôn ngữ",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "vi"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "old_password": {
                                        "type": "string"
                                    },
                                    "new_password": {
                                        "type": "string"
                                    },
                                    "confirm_password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            },
                            "examples": {
                                "Examples": {
                                    "summary": "Examples",
                                    "value": {
                                        "old_password": "123123",
                                        "new_password": "123456",
                                        "confirm_password": "123456",
                                        "code_verify": "123456"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Success."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/auth/cms/send-verify-code": {
            "post": {
                "tags": [
                    "Tài khoản CMS"
                ],
                "summary": "Gửi mã xác nhận để reset mật khẩu",
                "operationId": "6507efcb45d0fc64c4c03fad1b5d9bb1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "user@example.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Mã xác nhận đã được gửi"
                    },
                    "400": {
                        "description": "Lỗi validation"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}